refactor(fee-payer): isolated IP rate limiting
